Reduced the file size to 75% of the original .doc file.
The .doc file was 24 MB, the html generated from that was 41 MB, and after this
commit, the html has been reduced to 17 MB.

Removed MsoNormal class from <p> tags; made the <style> changes so that the same
style applies to <p> tags that was applied to <p> tags with class=MsoNormal.

Removed style definitions that were repeated in every <span> tag, and replaced
them with class attributes in those spans.

Created style elements for each of those classes, and applied the same style
definitions that were removed from the <span> tags.

The effect is that the document looks exactly the same in the browser, but is
just 41% in size as compared to the HTML version we started with!