Add a sample as index.html (generated file).
Since the full html file is about 13 MB, I decided to provide a sample
which gives people a taste of what's in the file, and then they can
decide to see full text of the project, if they wish to, by visiting the
link at the end of the sample.

The sample is generated by the generate_sample.sh script. So anytime
there's a change in the main file's head section (up to the line
containing 'SAMPLE PAGE END' text), then the sample should be
regenrated by simply executing the script.

Also added my name to the attributions list, with a link to the
project's web page to solicit collaboation from others interested in the
project.